backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
numpy>=1.26
Pillow>=10.0
python-dotenv==1.2.2
PyYAML==6.0.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
agno
openai
mcp[cli]
ddgs
# LiDAR GeoTIFF support — opzionale: installa solo se usi il layer LiDAR
# Richiede GDAL: Ubuntu: apt-get install libgdal-dev | macOS: brew install gdal
rasterio>=1.3
